+++ src/Interfaces/ManyRelationInterface.php
@@ -3,6 +3,7 @@
 namespace Swis\JsonApi\Client\Interfaces;
 
 use Swis\JsonApi\Client\Collection;
+use Swis\JsonApi\Client\Links;
 use Swis\JsonApi\Client\Meta;
 
 interface ManyRelationInterface
@@ -40,6 +41,16 @@ interface ManyRelationInterface
      * @return bool
      */
     public function shouldOmitIncluded(): bool;
+
+    /**
+     * @param \Swis\JsonApi\Client\Links|null $links
+     */
+    public function setLinks(Links $links = null);
+
+    /**
+     * @return \Swis\JsonApi\Client\Links|null
+     */
+    public function getLinks();
 
     /**
      * @param \Swis\JsonApi\Client\Meta|null $meta