+++ src/Interfaces/OneRelationInterface.php
@@ -2,6 +2,7 @@
 
 namespace Swis\JsonApi\Client\Interfaces;
 
+use Swis\JsonApi\Client\Links;
 use Swis\JsonApi\Client\Meta;
 
 interface OneRelationInterface
@@ -39,6 +40,16 @@ interface OneRelationInterface
      * @return bool
      */
     public function shouldOmitIncluded(): bool;
+
+    /**
+     * @param \Swis\JsonApi\Client\Links|null $links
+     */
+    public function setLinks(Links $links = null);
+
+    /**
+     * @return \Swis\JsonApi\Client\Links|null
+     */
+    public function getLinks();
 
     /**
      * @param \Swis\JsonApi\Client\Meta|null $meta